refactor(globe-scroll): move component to scroll-story folder
